feat(checksum): extract algorithm detection in a specific file
