+++ src/uucore/src/lib/features/checksum/mod.rs
@@ -20,16 +20,15 @@ use std::{
 use crate::{
     error::{set_exit_code, FromIo, UError, UResult, USimpleError},
     os_str_as_bytes, show, show_error, show_warning_caps,
-    sum::{
-        Blake2b, Blake3, Digest, DigestWriter, Md5, Sha1, Sha224, Sha256, Sha384, Sha3_224,
-        Sha3_256, Sha3_384, Sha3_512, Sha512, Shake128, Shake256, Sm3, BSD, CRC, SYSV,
-    },
+    sum::{Digest, DigestWriter},
     util_name,
 };
 use std::fmt::Write;
 use std::io::stdin;
 use std::io::BufRead;
 use thiserror::Error;
+
+pub mod algo;
 
 pub const ALGORITHM_OPTIONS_SYSV: &str = "sysv";
 pub const ALGORITHM_OPTIONS_BSD: &str = "bsd";
@@ -121,41 +120,6 @@ impl UError for ChecksumError {
     }
 }
 
-/// Creates a SHA3 hasher instance based on the specified bits argument.
-///
-/// # Returns
-///
-/// Returns a UResult of a tuple containing the algorithm name, the hasher instance, and
-/// the output length in bits or an Err if an unsupported output size is provided, or if
-/// the `--bits` flag is missing.
-pub fn create_sha3(bits: Option<usize>) -> UResult<HashAlgorithm> {
-    match bits {
-        Some(224) => Ok(HashAlgorithm {
-            name: "SHA3_224",
-            create_fn: Box::new(|| Box::new(Sha3_224::new())),
-            bits: 224,
-        }),
-        Some(256) => Ok(HashAlgorithm {
-            name: "SHA3_256",
-            create_fn: Box::new(|| Box::new(Sha3_256::new())),
-            bits: 256,
-        }),
-        Some(384) => Ok(HashAlgorithm {
-            name: "SHA3_384",
-            create_fn: Box::new(|| Box::new(Sha3_384::new())),
-            bits: 384,
-        }),
-        Some(512) => Ok(HashAlgorithm {
-            name: "SHA3_512",
-            create_fn: Box::new(|| Box::new(Sha3_512::new())),
-            bits: 512,
-        }),
-
-        Some(_) => Err(ChecksumError::InvalidOutputSizeForSha3.into()),
-        None => Err(ChecksumError::BitsRequiredForSha3.into()),
-    }
-}
-
 #[allow(clippy::comparison_chain)]
 fn cksum_output(res: &ChecksumResult, ignore_missing: bool, status: bool) {
     if res.bad_format == 1 {
@@ -177,105 +141,6 @@ fn cksum_output(res: &ChecksumResult, ignore_missing: bool, status: bool) {
         } else if res.failed_open_file > 1 {
             show_warning_caps!("{} listed files could not be read", res.failed_open_file);
         }
-    }
-}
-
-pub fn detect_algo(algo: &str, length: Option<usize>) -> UResult<HashAlgorithm> {
-    match algo {
-        ALGORITHM_OPTIONS_SYSV => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_SYSV,
-            create_fn: Box::new(|| Box::new(SYSV::new())),
-            bits: 512,
-        }),
-        ALGORITHM_OPTIONS_BSD => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_BSD,
-            create_fn: Box::new(|| Box::new(BSD::new())),
-            bits: 1024,
-        }),
-        ALGORITHM_OPTIONS_CRC => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_CRC,
-            create_fn: Box::new(|| Box::new(CRC::new())),
-            bits: 256,
-        }),
-        ALGORITHM_OPTIONS_MD5 | "md5sum" => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_MD5,
-            create_fn: Box::new(|| Box::new(Md5::new())),
-            bits: 128,
-        }),
-        ALGORITHM_OPTIONS_SHA1 | "sha1sum" => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_SHA1,
-            create_fn: Box::new(|| Box::new(Sha1::new())),
-            bits: 160,
-        }),
-        ALGORITHM_OPTIONS_SHA224 | "sha224sum" => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_SHA224,
-            create_fn: Box::new(|| Box::new(Sha224::new())),
-            bits: 224,
-        }),
-        ALGORITHM_OPTIONS_SHA256 | "sha256sum" => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_SHA256,
-            create_fn: Box::new(|| Box::new(Sha256::new())),
-            bits: 256,
-        }),
-        ALGORITHM_OPTIONS_SHA384 | "sha384sum" => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_SHA384,
-            create_fn: Box::new(|| Box::new(Sha384::new())),
-            bits: 384,
-        }),
-        ALGORITHM_OPTIONS_SHA512 | "sha512sum" => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_SHA512,
-            create_fn: Box::new(|| Box::new(Sha512::new())),
-            bits: 512,
-        }),
-        ALGORITHM_OPTIONS_BLAKE2B | "b2sum" => {
-            // Set default length to 512 if None
-            let bits = length.unwrap_or(512);
-            if bits == 512 {
-                Ok(HashAlgorithm {
-                    name: ALGORITHM_OPTIONS_BLAKE2B,
-                    create_fn: Box::new(move || Box::new(Blake2b::new())),
-                    bits: 512,
-                })
-            } else {
-                Ok(HashAlgorithm {
-                    name: ALGORITHM_OPTIONS_BLAKE2B,
-                    create_fn: Box::new(move || Box::new(Blake2b::with_output_bytes(bits))),
-                    bits,
-                })
-            }
-        }
-        ALGORITHM_OPTIONS_BLAKE3 | "b3sum" => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_BLAKE3,
-            create_fn: Box::new(|| Box::new(Blake3::new())),
-            bits: 256,
-        }),
-        ALGORITHM_OPTIONS_SM3 => Ok(HashAlgorithm {
-            name: ALGORITHM_OPTIONS_SM3,
-            create_fn: Box::new(|| Box::new(Sm3::new())),
-            bits: 512,
-        }),
-        ALGORITHM_OPTIONS_SHAKE128 | "shake128sum" => {
-            let bits =
-                length.ok_or_else(|| USimpleError::new(1, "--bits required for SHAKE128"))?;
-            Ok(HashAlgorithm {
-                name: ALGORITHM_OPTIONS_SHAKE128,
-                create_fn: Box::new(|| Box::new(Shake128::new())),
-                bits,
-            })
-        }
-        ALGORITHM_OPTIONS_SHAKE256 | "shake256sum" => {
-            let bits =
-                length.ok_or_else(|| USimpleError::new(1, "--bits required for SHAKE256"))?;
-            Ok(HashAlgorithm {
-                name: ALGORITHM_OPTIONS_SHAKE256,
-                create_fn: Box::new(|| Box::new(Shake256::new())),
-                bits,
-            })
-        }
-        //ALGORITHM_OPTIONS_SHA3 | "sha3" => (
-        _ if algo.starts_with("sha3") => create_sha3(length),
-
-        _ => Err(ChecksumError::UnknownAlgorithm.into()),
     }
 }
 
@@ -601,7 +466,7 @@ where
                     properly_formatted = false;
                     continue;
                 }
-                let mut algo = detect_algo(&algo_name, length)?;
+                let mut algo = algo::detect_algo(&algo_name, length)?;
 
                 let (filename_to_check_unescaped, prefix) = unescape_filename(filename_to_check);
 
@@ -850,76 +715,6 @@ mod tests {
         assert!(calculate_blake2b_length(520).is_err());
         assert_eq!(calculate_blake2b_length(512).unwrap(), None);
         assert_eq!(calculate_blake2b_length(256).unwrap(), Some(32));
-    }
-
-    #[test]
-    fn test_detect_algo() {
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SYSV, None).unwrap().name,
-            ALGORITHM_OPTIONS_SYSV
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_BSD, None).unwrap().name,
-            ALGORITHM_OPTIONS_BSD
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_CRC, None).unwrap().name,
-            ALGORITHM_OPTIONS_CRC
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_MD5, None).unwrap().name,
-            ALGORITHM_OPTIONS_MD5
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SHA1, None).unwrap().name,
-            ALGORITHM_OPTIONS_SHA1
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SHA224, None).unwrap().name,
-            ALGORITHM_OPTIONS_SHA224
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SHA256, None).unwrap().name,
-            ALGORITHM_OPTIONS_SHA256
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SHA384, None).unwrap().name,
-            ALGORITHM_OPTIONS_SHA384
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SHA512, None).unwrap().name,
-            ALGORITHM_OPTIONS_SHA512
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_BLAKE2B, None).unwrap().name,
-            ALGORITHM_OPTIONS_BLAKE2B
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_BLAKE3, None).unwrap().name,
-            ALGORITHM_OPTIONS_BLAKE3
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SM3, None).unwrap().name,
-            ALGORITHM_OPTIONS_SM3
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SHAKE128, Some(128))
-                .unwrap()
-                .name,
-            ALGORITHM_OPTIONS_SHAKE128
-        );
-        assert_eq!(
-            detect_algo(ALGORITHM_OPTIONS_SHAKE256, Some(256))
-                .unwrap()
-                .name,
-            ALGORITHM_OPTIONS_SHAKE256
-        );
-        assert_eq!(detect_algo("sha3_224", Some(224)).unwrap().name, "SHA3_224");
-        assert_eq!(detect_algo("sha3_256", Some(256)).unwrap().name, "SHA3_256");
-        assert_eq!(detect_algo("sha3_384", Some(384)).unwrap().name, "SHA3_384");
-        assert_eq!(detect_algo("sha3_512", Some(512)).unwrap().name, "SHA3_512");
-
-        assert!(detect_algo("sha3_512", None).is_err());
     }
 
     #[test]
